:construction: Moving method to parent class
